# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Phonebook App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "The **Phonebook App** is a RESTful API that allows users to securely manage their personal and professional contacts. This documentation provides a detailed overview of all available endpoints, request/response formats, and authentication requirements."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://hw2-mongodb-07bu.onrender.com"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "description": "Creates a new user account. This endpoint requires a **name**, **email**, and **password**.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully registered user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully registered a user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/userCreatedResponse"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "userRegistrationRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "description": "Full name of the user. Must be 3-30 characters.",
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "Alice Woods"
          },
          "email": {
            "description": "User's unique email address. Each email can be registered only ones.",
            "type": "string",
            "example": "alicewoods@example.com"
          },
          "password": {
            "description": "User's password. Must be kept confidential.",
            "type": "string",
            "example": "Nk45Yvs!13t"
          }
        }
      },
      "userCreatedResponse": {
        "type": "object",
        "required": [
          "name",
          "email",
          "_id",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "name": {
            "description": "Full name of the user.",
            "type": "string",
            "minLength": 3,
            "maxLength": 30,
            "example": "Alice Woods"
          },
          "email": {
            "description": "The user's unique email address.",
            "type": "string",
            "example": "awoods@mail.com"
          },
          "_id": {
            "description": "Unique identifier for the user, automatically generated by the database.",
            "type": "string",
            "example": "687e27369175a01d77ac83eb"
          },
          "createdAt": {
            "description": "Date and time when the user account was created (ISO 8601 format, UTC), automatically generated by the database.",
            "type": "string",
            "format": "date-time",
            "example": "2025-07-18T11:00:15.361Z"
          },
          "updatedAt": {
            "description": "Date and time when the user's information was last updated (ISO 8601 format, UTC), automatically generated by the database.",
            "type": "string",
            "format": "date-time",
            "example": "2025-08-11T10:52:15.353Z"
          }
        }
      }
    },
    "responses": {
...
      "409": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "Conflict"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email is already in use"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}